+++ packages/react-icons-atomic-webpack-loader/test/src/aliased-indirect-reexports.js
@@ -1,0 +1,3 @@
+import { AddFilled as MyAdd, ArrowLeftRegular as MyArrow } from '@fluentui/react-icons';
+
+export { MyAdd, MyArrow };