+++ packages/react-icons-atomic-webpack-loader/test/src/renamed-indirect-reexports.js
@@ -1,0 +1,3 @@
+import { AddFilled, ArrowLeftRegular } from '@fluentui/react-icons';
+
+export { AddFilled as MyAdd, ArrowLeftRegular as MyArrow };