+++ packages/react-icons-atomic-webpack-loader/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2 (2026-05-18)
+
+### 🩹 Fixes
+
+- **react-icons-atomic-webpack-loader:** skip indirect re-exports to avoid duplicate-export build failures ([#1082](https://github.com/microsoft/fluentui-system-icons/pull/1082))
+- **release:** prevent cross-group version cascade to standalone packages ([#1072](https://github.com/microsoft/fluentui-system-icons/pull/1072))
+
 ## 0.0.1 (2026-04-17)
 
 ### 🚀 Features